Always store deserialized table Rows as LongNode
- No longer lose precision when round-tripping long values in
  Schema.INTEGER column
- Add Row.getLong to retrieve long values
- fixes Jackson serialization round-trip JsonNode subclass mismatch by postprocessing table rows and excluded rows on deserialization according to schema

+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/answers/Schema.java
@@ -45,7 +45,7 @@ public class Schema {
           .put("FileLines", getClassString(FileLines.class))
           .put("Flow", getClassString(Flow.class))
           .put("FlowTrace", getClassString(FlowTrace.class))
-          .put("Integer", getClassString(Integer.class))
+          .put("Integer", getClassString(Long.class))
           .put("Interface", getClassString(NodeInterfacePair.class))
           .put("Ip", getClassString(Ip.class))
           .put("Issue", getClassString(Issue.class))
@@ -151,7 +151,7 @@ public class Schema {
 
   /** Whether this Schema object is Integer-based (base, list, or set) */
   public boolean isIntBased() {
-    return _baseType.equals(Integer.class);
+    return _baseType.equals(Long.class);
   }
 
   @Override
