+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/answers/AnswerMetadataUtilTest.java
@@ -36,7 +36,7 @@ public class AnswerMetadataUtilTest {
   public void testComputeAnswerMetadata() throws IOException {
     String columnName = "col";
     String issueColumnName = "colIssue";
-    int value = 5;
+    long value = 5L;
     String major = "major";
     String minor = "minor";
     int severity = 1;
@@ -64,7 +64,7 @@ public class AnswerMetadataUtilTest {
                                 columnName,
                                 ImmutableMap.of(Aggregation.MAX, value),
                                 issueColumnName,
-                                ImmutableMap.of(Aggregation.MAX, severity)))
+                                ImmutableMap.of(Aggregation.MAX, (long) severity)))
                         .setMajorIssueConfigs(
                             ImmutableMap.of(
                                 major,
@@ -115,7 +115,7 @@ public class AnswerMetadataUtilTest {
   @Test
   public void testComputeColumnAggregationMax() {
     String columnName = "col";
-    int value = 5;
+    long value = 5L;
 
     TableAnswerElement table =
         new TableAnswerElement(
@@ -134,7 +134,7 @@ public class AnswerMetadataUtilTest {
   @Test
   public void testComputeColumnAggregations() {
     String columnName = "col";
-    int value = 5;
+    long value = 5L;
 
     TableAnswerElement table =
         new TableAnswerElement(
@@ -198,7 +198,7 @@ public class AnswerMetadataUtilTest {
   @Test
   public void testComputeColumnMaxOneRowInteger() {
     String columnName = "col";
-    int value = 5;
+    long value = 5;
 
     TableAnswerElement table =
         new TableAnswerElement(
@@ -223,14 +223,15 @@ public class AnswerMetadataUtilTest {
                     new DisplayHints().getTextDesc()))
             .addRow(Row.of(columnName, value));
 
-    assertThat(AnswerMetadataUtil.computeColumnMax(table, columnName, _logger), equalTo(severity));
+    assertThat(
+        AnswerMetadataUtil.computeColumnMax(table, columnName, _logger), equalTo((long) severity));
   }
 
   @Test
   public void testComputeColumnMaxTwoRows() {
     String columnName = "col";
-    int value1 = 5;
-    int value2 = 10;
+    long value1 = 5L;
+    long value2 = 10L;
 
     TableAnswerElement table =
         new TableAnswerElement(
